Move internal Modifier.expandable to common
