+++ compose/material/material/src/androidMain/kotlin/androidx/compose/material/ExposedDropdownMenu.kt
@@ -226,30 +226,6 @@ internal actual fun ExposedDropdownMenuBoxScope.ExposedDropdownMenuDefaultImpl(
     }
 }
 
-private fun Modifier.expandable(
-    onExpandedChange: () -> Unit,
-    menuLabel: String
-) = composed {
-    val currentOnExpandedChange by rememberUpdatedState(onExpandedChange)
-    pointerInput(Unit) {
-        awaitEachGesture {
-            // Must be PointerEventPass.Initial to observe events before the text field consumes them
-            // in the Main pass
-            awaitFirstDown(pass = PointerEventPass.Initial)
-            val upEvent = waitForUpOrCancellation(pass = PointerEventPass.Initial)
-            if (upEvent != null) {
-                currentOnExpandedChange()
-            }
-        }
-    }.semantics {
-        contentDescription = menuLabel // this should be a localised string
-        onClick {
-            currentOnExpandedChange()
-            true
-        }
-    }
-}
-
 private fun updateHeight(
     view: View,
     coordinates: LayoutCoordinates?,
