+++ compose/material3/material3/src/skikoMain/kotlin/androidx/compose/material3/ExposedDropdownMenu.skiko.kt
@@ -151,33 +151,3 @@ internal actual fun ExposedDropdownMenuBoxScope.ExposedDropdownMenuDefaultImpl(
         content = content
     )
 }
-
-@Suppress("ComposableModifierFactory")
-@Composable
-private fun Modifier.expandable(
-    expanded: Boolean,
-    onExpandedChange: () -> Unit,
-    menuDescription: String = getString(Strings.ExposedDropdownMenu),
-    expandedDescription: String = getString(Strings.MenuExpanded),
-    collapsedDescription: String = getString(Strings.MenuCollapsed),
-) = composed {
-    val currentOnExpandedChange by rememberUpdatedState(onExpandedChange)
-    pointerInput(Unit) {
-        awaitEachGesture {
-            // Must be PointerEventPass.Initial to observe events before the text field consumes them
-            // in the Main pass
-            awaitFirstDown(pass = PointerEventPass.Initial)
-            val upEvent = waitForUpOrCancellation(pass = PointerEventPass.Initial)
-            if (upEvent != null) {
-                currentOnExpandedChange()
-            }
-        }
-    }.semantics {
-        stateDescription = if (expanded) expandedDescription else collapsedDescription
-        contentDescription = menuDescription
-        onClick {
-            currentOnExpandedChange()
-            true
-        }
-    }
-}
